Move renderResult under lib-components, remove UnwrapResult
