+++ frontend/src/lib-components/async-rendering.tsx
@@ -1,0 +1,15 @@
+import { Result } from 'lib-common/api'
+import React from 'react'
+import { SpinnerSegment } from './atoms/state/Spinner'
+import ErrorSegment from './atoms/state/ErrorSegment'
+
+export function renderResult<T>(
+  result: Result<T>,
+  renderer: (data: T) => React.ReactNode
+) {
+  if (result.isLoading) return <SpinnerSegment />
+
+  if (result.isFailure) return <ErrorSegment />
+
+  return renderer(result.value)
+}
